refactor: Change folder structure & styles for full page title
