+++ src/app/app-routing.module.ts
@@ -1,6 +1,5 @@
 import { NgModule } from '@angular/core';
-import { Routes, RouterModule } from '@angular/router';
-import { SpeakersComponent } from './speakers/speakers.component';
+import { RouterModule, Routes } from '@angular/router';
 
 const routes: Routes = [
   {
