+++ src/app/speakers/core/speaker-photo.pipe.spec.ts
