+++ src/app/speakers/core/speaker-photo.pipe.ts
