+++ src/app/speakers/core/speaker.model.ts
@@ -6,6 +6,7 @@ export class Speaker {
     public photo: string = null,
     public twitter: string = null,
     public github: string = null,
+    public website: string = null,
   ) {
     if (this.photo === null) {
       this.photo = `https://picsum.photos/600/400?random&image=${Math.round(Math.random() * 100)}`;
