+++ src/app/speakers/speakers-bio/speakers-bio.component.ts
@@ -1,7 +1,7 @@
 import { ChangeDetectionStrategy, Component, Inject } from '@angular/core';
 import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material';
 
-import { Speaker } from '../speaker.model';
+import { Speaker } from '../core/speaker.model';
 
 @Component({
   selector: 'app-speakers-bio',
