+++ src/app/speakers/speakers-list/speakers-list.component.html
@@ -1,11 +1,13 @@
-<div id='title'>
-  <div>Speakers</div>
-</div>
+<div id='firstScreen'>
+  <div id='title'>
+    <div>Speakers</div>
+  </div>
 
-<div id='scrollDown'>
-  <mat-icon>expand_more</mat-icon>
-  <div>
-    Scroll Down
+  <div id='scrollDown'>
+    <mat-icon>expand_more</mat-icon>
+    <div>
+      Scroll Down
+    </div>
   </div>
 </div>
