+++ src/app/speakers/speakers-list/speakers-list.component.ts
@@ -1,16 +1,16 @@
 import { ChangeDetectionStrategy, Component } from '@angular/core';
 import { MatDialog } from '@angular/material';
 
-import { Speaker } from './speaker.model';
-import { SpeakersBioComponent } from './speakers-bio/speakers-bio.component';
+import { Speaker } from '../core/speaker.model';
+import { SpeakersBioComponent } from '../speakers-bio/speakers-bio.component';
 
 @Component({
-  selector: 'app-speakers',
-  templateUrl: './speakers.component.html',
-  styleUrls: ['./speakers.component.scss'],
+  selector: 'app-speakers-list',
+  templateUrl: './speakers-list.component.html',
+  styleUrls: ['./speakers-list.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class SpeakersComponent {
+export class SpeakersListComponent {
   public speakers: Speaker[];
 
   constructor(public dialog: MatDialog) {
@@ -22,6 +22,7 @@ export class SpeakersComponent {
         'https://secure.gravatar.com/avatar/357efbf6f7a1495e5854cb47728fa523?s=64',
         'https://twitter.com/luchillo17',
         'https://github.com/luchillo17',
+        'https://www.instagram.com/luchillo24/',
       ),
       ...new Array(5).fill(1).map(() => new Speaker()),
     ];
