+++ src/app/speakers/speakers.module.ts
@@ -3,14 +3,14 @@ import { NgModule } from '@angular/core';
 import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
 
 import { MaterialModule } from '../material/material.module';
-import { SpeakerPhotoUrlPipe } from './speaker-photo.pipe';
+import { SpeakerPhotoUrlPipe } from './core/speaker-photo.pipe';
 import { SpeakersBioComponent } from './speakers-bio/speakers-bio.component';
+import { SpeakersListComponent } from './speakers-list/speakers-list.component';
 import { SpeakersRoutingModule } from './speakers-routing.module';
-import { SpeakersComponent } from './speakers.component';
 
 @NgModule({
   imports: [CommonModule, FontAwesomeModule, MaterialModule, SpeakersRoutingModule],
-  declarations: [SpeakerPhotoUrlPipe, SpeakersBioComponent, SpeakersComponent],
+  declarations: [SpeakerPhotoUrlPipe, SpeakersBioComponent, SpeakersListComponent],
   entryComponents: [SpeakersBioComponent],
 })
 export class SpeakersModule {}
